refactored Progress Viewer files
