+++ src/UI/TrainingViewer.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { useState, useEffect } from "react";
 import axios from "axios";
-import FileDropdown from "./ViewerComponents/FileDropdown";
-import ProgressFileReader from "./ViewerComponents/ProgressFileReader";
+import FileDropdown from "../components/progress-viewer/file-dropdown";
+import ProgressFileReader from "../components/progress-viewer/progress-file-reader";
 
 
 const TrainingViewer = ({}) => {
